Fix rebase issue and consolidate eagle3_attention tests
